requirements.txt
apipkg==1.5
appnope==0.1.0
attrs==19.3.0
backcall==0.2.0
beautifulsoup4==4.9.1
boto==2.49.0
brython==3.8.9
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
commonmark==0.9.1
coverage==5.1
cryptography==2.9.2
decorator==4.4.2
execnet==1.7.1
flake8==3.8.3
idna==2.9
importlib-metadata==1.7.0
ipdb==0.13.3
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.1
mccabe==0.6.1
more-itertools==8.4.0
nose==1.3.7
numpy==1.19.0
packaging==20.4
pandas==1.0.5
parameterized==0.7.4
parso==0.7.0
pdfminer.six==20200517
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
pprintpp==0.4.0
prompt-toolkit==3.0.5
ptyprocess==0.6.0
py==1.8.1
pycodestyle==2.6.0
pycparser==2.20
pycryptodome==3.9.8
pyflakes==2.2.0
Pygments==2.6.1
PyMySQL==0.9.3
pyOpenSSL==19.1.0
pyotp==2.3.0
pyparsing==2.4.7
pytest==5.4.3
pytest-cov==2.10.0
pytest-forked==1.2.0
pytest-html==2.0.1
pytest-metadata==1.10.0
pytest-ordering==0.6
pytest-rerunfailures==9.0
pytest-xdist==1.32.0
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
rich==2.2.6
selenium==3.141.0
seleniumbase==1.41.0
setuptools-scm==4.1.2
six==1.15.0
sortedcontainers==2.2.2
soupsieve==2.0.1
traitlets==4.3.3
typing-extensions==3.7.4.2
urllib3==1.25.9
wcwidth==0.2.5
zipp==3.1.0

html2text~=2020.1.16
google-api-python-client~=1.9.3